fixed weapon range serialization
Former-commit-id: ec6ab8a5bbd3e930b7667577ea8922ceb4e7827e

+++ SpaceComponentSystem/Components/Util/ShipInfo.cs
@@ -254,7 +254,7 @@ namespace Space.ComponentSystem.Components
         /// <summary>
         /// The distance our highest range weapon can shoot.
         /// </summary>
-        public double WeaponRange { get; internal set; }
+        public float WeaponRange { get; internal set; }
 
         #endregion
